10/27 comple Signed-off-by: HoSeon <a01073591105@gmail.com>

+++ MiniProject/src/main/java/com/java96/mapper/UploadMapper.java
@@ -18,7 +18,7 @@ public interface UploadMapper {
 	
 	
 	@Select("select * from tbl_testattach where tno= #{tno}")
-	public List<UploadDTO> viewAttach(Long tno);
+	public List<UploadDTO> viewAttach(int tno);
